fix: separate CSR, RegistrationInfo and EntityInfo
