+++ search-service/src/test/kotlin/com/egm/stellio/search/csr/CsrUtils.kt
@@ -1,8 +1,10 @@
 package com.egm.stellio.search.csr
 
 import com.egm.stellio.search.csr.model.ContextSourceRegistration
+import com.egm.stellio.search.csr.model.EntityInfo
 import com.egm.stellio.search.csr.model.Mode
 import com.egm.stellio.search.csr.model.Operation
+import com.egm.stellio.search.csr.model.RegistrationInfo
 import com.egm.stellio.shared.util.APIARY_TYPE
 import com.egm.stellio.shared.util.ngsiLdDateTime
 import com.egm.stellio.shared.util.toUri
@@ -12,9 +14,9 @@ object CsrUtils {
     fun gimmeRawCSR(
         id: URI = "urn:ngsi-ld:ContextSourceRegistration:test".toUri(),
         endpoint: URI = "http://localhost:8089".toUri(),
-        information: List<ContextSourceRegistration.RegistrationInfo> = listOf(
-            ContextSourceRegistration.RegistrationInfo(
-                listOf(ContextSourceRegistration.EntityInfo(types = listOf(APIARY_TYPE)))
+        information: List<RegistrationInfo> = listOf(
+            RegistrationInfo(
+                listOf(EntityInfo(types = listOf(APIARY_TYPE)))
             )
         ),
         operations: List<Operation> = listOf(Operation.FEDERATION_OPS),
