+++ search-service/src/test/kotlin/com/egm/stellio/search/csr/model/ContextSourceRegistrationTests.kt
@@ -1,16 +1,12 @@
 package com.egm.stellio.search.csr.model
 
 import com.egm.stellio.search.csr.CsrUtils.gimmeRawCSR
-import com.egm.stellio.search.csr.model.ContextSourceRegistration.RegistrationInfo
 import com.egm.stellio.shared.model.BadRequestDataException
 import com.egm.stellio.shared.util.APIC_COMPOUND_CONTEXT
-import com.egm.stellio.shared.util.APIC_COMPOUND_CONTEXTS
 import com.egm.stellio.shared.util.BEEHIVE_TYPE
 import com.egm.stellio.shared.util.JsonLdUtils.NGSILD_CSR_TERM
-import com.egm.stellio.shared.util.MANAGED_BY_COMPACT_RELATIONSHIP
 import com.egm.stellio.shared.util.MANAGED_BY_RELATIONSHIP
 import com.egm.stellio.shared.util.NGSILD_NAME_PROPERTY
-import com.egm.stellio.shared.util.NGSILD_NAME_TERM
 import com.egm.stellio.shared.util.expandJsonLdEntity
 import com.egm.stellio.shared.util.shouldFailWith
 import com.egm.stellio.shared.util.shouldSucceed
@@ -18,8 +14,6 @@ import com.egm.stellio.shared.util.shouldSucceedAndResult
 import com.egm.stellio.shared.util.toUri
 import kotlinx.coroutines.test.runTest
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Test
 import java.net.URI
 
@@ -148,7 +142,7 @@ class ContextSourceRegistrationTests {
             ids = setOf(entity.id),
             types = entity.types.toSet()
         )
-        val entityInfo = ContextSourceRegistration.EntityInfo(entity.id, types = entity.types)
+        val entityInfo = EntityInfo(entity.id, types = entity.types)
         val information = RegistrationInfo(
             entities = listOf(entityInfo),
             propertyNames = listOf(NGSILD_NAME_PROPERTY),
@@ -172,7 +166,7 @@ class ContextSourceRegistrationTests {
             ids = setOf(entity.id),
             types = entity.types.toSet()
         )
-        val nonMatchingEntityInfo = ContextSourceRegistration.EntityInfo(types = listOf(BEEHIVE_TYPE))
+        val nonMatchingEntityInfo = EntityInfo(types = listOf(BEEHIVE_TYPE))
         val nonMatchingInformation = RegistrationInfo(
             entities = listOf(nonMatchingEntityInfo),
             propertyNames = listOf(NGSILD_NAME_PROPERTY),
@@ -189,94 +183,18 @@ class ContextSourceRegistrationTests {
     }
 
     @Test
-    fun `getAttributeNames should merge propertyNames and relationshipNames`() = runTest {
-        val information = RegistrationInfo(
-            propertyNames = listOf(NGSILD_NAME_PROPERTY),
-            relationshipNames = listOf(MANAGED_BY_RELATIONSHIP)
-        )
-
-        val attrs = information.getAttributeNames()
-        assertThat(attrs).hasSize(2)
-        assertThat(attrs).contains(NGSILD_NAME_PROPERTY, MANAGED_BY_RELATIONSHIP)
-    }
-
-    @Test
-    fun `getAttributeNames should keep relationshipNames if propertyNames is null`() = runTest {
-        val information = RegistrationInfo(
-            propertyNames = null,
-            relationshipNames = listOf(MANAGED_BY_RELATIONSHIP)
-        )
-
-        val attrs = information.getAttributeNames()
-        assertEquals(setOf(MANAGED_BY_RELATIONSHIP), attrs)
-    }
-
-    @Test
-    fun `getAttributeNames should keep propertyNames if relationshipNames is null`() = runTest {
-        val information = RegistrationInfo(
-            propertyNames = listOf(NGSILD_NAME_PROPERTY),
-            relationshipNames = null
-        )
-
-        val attrs = information.getAttributeNames()
-        assertEquals(setOf(NGSILD_NAME_PROPERTY), attrs)
-    }
-
-    @Test
-    fun `getAttributeNames should return null if propertyNames and relationshipNames are null`() = runTest {
-        val information = RegistrationInfo(
-            propertyNames = null,
-            relationshipNames = null
-        )
-
-        val attrs = information.getAttributeNames()
-        assertNull(attrs)
-    }
-
-    @Test
-    fun `computeAttrsQueryParam should intersect the csf and the registration attributes`() = runTest {
-        val registrationInfo = RegistrationInfo(
-            propertyNames = listOf(MANAGED_BY_RELATIONSHIP, NGSILD_NAME_PROPERTY)
-        )
-        val csrFilters = CSRFilters(attrs = setOf(NGSILD_NAME_PROPERTY))
-
-        val attrs = registrationInfo.computeAttrsQueryParam(csrFilters, APIC_COMPOUND_CONTEXTS)
-        assertEquals(NGSILD_NAME_TERM, attrs)
-    }
-
-    @Test
-    fun `computeAttrsQueryParam should return the registration attributes if the csf is empty`() = runTest {
-        val registrationInfo = RegistrationInfo(
-            propertyNames = listOf(MANAGED_BY_RELATIONSHIP, NGSILD_NAME_PROPERTY)
-        )
-        val csrFilters = CSRFilters()
-
-        val attrs = registrationInfo.computeAttrsQueryParam(csrFilters, APIC_COMPOUND_CONTEXTS)
-        assertEquals("$MANAGED_BY_COMPACT_RELATIONSHIP,$NGSILD_NAME_TERM", attrs)
-    }
-
-    @Test
-    fun `computeAttrsQueryParam should return the csf attributes if the registration have no attributes`() = runTest {
-        val registrationInfo = RegistrationInfo()
-        val csrFilters = CSRFilters(attrs = setOf(NGSILD_NAME_PROPERTY))
-
-        val attrs = registrationInfo.computeAttrsQueryParam(csrFilters, APIC_COMPOUND_CONTEXTS)
-        assertEquals(NGSILD_NAME_TERM, attrs)
-    }
-
-    @Test
     fun `toSingleEntityInfoCSRList should return a list of csr with one entityInfo each`() = runTest {
         val registrationInformations = listOf(
             RegistrationInfo(
                 entities = listOf(
-                    ContextSourceRegistration.EntityInfo(id = "urn:1".toUri(), types = listOf(BEEHIVE_TYPE)),
-                    ContextSourceRegistration.EntityInfo(id = "urn:2".toUri(), types = listOf(BEEHIVE_TYPE))
+                    EntityInfo(id = "urn:1".toUri(), types = listOf(BEEHIVE_TYPE)),
+                    EntityInfo(id = "urn:2".toUri(), types = listOf(BEEHIVE_TYPE))
                 )
             ),
             RegistrationInfo(
                 entities = listOf(
-                    ContextSourceRegistration.EntityInfo(id = "urn:3".toUri(), types = listOf(BEEHIVE_TYPE)),
-                    ContextSourceRegistration.EntityInfo(id = "urn:4".toUri(), types = listOf(BEEHIVE_TYPE))
+                    EntityInfo(id = "urn:3".toUri(), types = listOf(BEEHIVE_TYPE)),
+                    EntityInfo(id = "urn:4".toUri(), types = listOf(BEEHIVE_TYPE))
                 )
             )
         )
@@ -293,14 +211,14 @@ class ContextSourceRegistrationTests {
         val registrationInformations = listOf(
             RegistrationInfo(
                 entities = listOf(
-                    ContextSourceRegistration.EntityInfo(id = "urn:1".toUri(), types = listOf(BEEHIVE_TYPE)),
-                    ContextSourceRegistration.EntityInfo(id = "urn:2".toUri(), types = listOf(BEEHIVE_TYPE))
+                    EntityInfo(id = "urn:1".toUri(), types = listOf(BEEHIVE_TYPE)),
+                    EntityInfo(id = "urn:2".toUri(), types = listOf(BEEHIVE_TYPE))
                 )
             ),
             RegistrationInfo(
                 entities = listOf(
-                    ContextSourceRegistration.EntityInfo(id = "urn:3".toUri(), types = listOf(BEEHIVE_TYPE)),
-                    ContextSourceRegistration.EntityInfo(id = "urn:4".toUri(), types = listOf(BEEHIVE_TYPE))
+                    EntityInfo(id = "urn:3".toUri(), types = listOf(BEEHIVE_TYPE)),
+                    EntityInfo(id = "urn:4".toUri(), types = listOf(BEEHIVE_TYPE))
                 )
             )
         )
